feat:update more user api

+++ backend/src/main/java/com/repairhub/management/common/dto/PageResponse.java
@@ -13,5 +13,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class PageResponse<T> {
     private List<T> list;
-    private Long total;
+    private int total;
 }
